chore: move tests to correct directory
