+++ src/test/java/uk/gov/hmcts/reform/preapi/tasks/migration/MigrateExclusionsTest.java
@@ -1,4 +1,4 @@
-package uk.gov.hmcts.reform.preapi.tasks.batch;
+package uk.gov.hmcts.reform.preapi.tasks.migration;
 
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
@@ -19,7 +19,6 @@ import uk.gov.hmcts.reform.preapi.dto.base.BaseAppAccessDTO;
 import uk.gov.hmcts.reform.preapi.security.authentication.UserAuthentication;
 import uk.gov.hmcts.reform.preapi.security.service.UserAuthenticationService;
 import uk.gov.hmcts.reform.preapi.services.UserService;
-import uk.gov.hmcts.reform.preapi.tasks.migration.FetchXML;
 
 import java.util.Optional;
 import java.util.Set;
@@ -32,13 +31,13 @@ import static org.mockito.Mockito.times;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 
-public class FetchXMLTest {
+public class MigrateExclusionsTest {
 
     private static UserService userService;
     private static UserAuthenticationService userAuthenticationService;
     private static JobLauncher jobLauncher;
     private static LoggingService loggingService;
-    private static Job fetchXmlJob;
+    private static Job processExclusionsJob;
 
     private static final String CRON_USER_EMAIL = "test@test.com";
 
@@ -59,30 +58,29 @@ public class FetchXMLTest {
 
         jobLauncher = mock(JobLauncher.class);
         loggingService = mock(LoggingService.class);
-        fetchXmlJob = mock(Job.class);
+        processExclusionsJob = mock(Job.class);
     }
 
-    @DisplayName("Test Fetch XML")
+    @DisplayName("Test Migrating Exclusions")
     @Test
     public void testRun() throws JobInstanceAlreadyCompleteException,
         JobExecutionAlreadyRunningException,
         JobParametersInvalidException,
         JobRestartException {
 
-        var fetchXML = new FetchXML(userService,
-                                    userAuthenticationService,
-                                    CRON_USER_EMAIL,
-                                    jobLauncher,
-                                    loggingService,
-                                    false,
-                                    false,
-                                    MigrationType.FULL.name(),
-                                    fetchXmlJob);
-        fetchXML.run();
+        var migrateExclusions = new MigrateExclusions(userService,
+                                                      userAuthenticationService,
+                                                      CRON_USER_EMAIL,
+                                                      jobLauncher,
+                                                      loggingService,
+                                                      false,
+                                                      false,
+                                                      processExclusionsJob);
+        migrateExclusions.run();
 
         ArgumentCaptor<JobParameters> jobParameters = ArgumentCaptor.forClass(JobParameters.class);
 
-        verify(jobLauncher, times(1)).run(eq(fetchXmlJob), jobParameters.capture());
+        verify(jobLauncher, times(1)).run(eq(processExclusionsJob), jobParameters.capture());
 
         Assertions.assertEquals(String.valueOf(false),
                                 jobParameters.getValue().getString("debug"));
@@ -90,63 +88,31 @@ public class FetchXMLTest {
         Assertions.assertEquals(MigrationType.FULL.name(),
                                 jobParameters.getValue().getString("migrationType"));
 
-        verify(loggingService, times(1)).logInfo("Successfully completed Fetch XML batch job");
+        verify(loggingService, times(1)).logInfo("Successfully completed Process Exclusions batch job");
     }
 
-    @DisplayName("Test Fetch XML Second Type")
-    @Test
-    public void testRunSecondType() throws JobInstanceAlreadyCompleteException,
-        JobExecutionAlreadyRunningException,
-        JobParametersInvalidException,
-        JobRestartException {
-
-        var fetchXML = new FetchXML(userService,
-                                    userAuthenticationService,
-                                    CRON_USER_EMAIL,
-                                    jobLauncher,
-                                    loggingService,
-                                    false,
-                                    false,
-                                    MigrationType.DELTA.name(),
-                                    fetchXmlJob);
-        fetchXML.run();
-
-        ArgumentCaptor<JobParameters> jobParameters = ArgumentCaptor.forClass(JobParameters.class);
-
-        verify(jobLauncher, times(1)).run(eq(fetchXmlJob), jobParameters.capture());
-
-        Assertions.assertEquals(String.valueOf(false),
-                                jobParameters.getValue().getString("debug"));
-
-        Assertions.assertEquals(MigrationType.DELTA.name(),
-                                jobParameters.getValue().getString("migrationType"));
-
-        verify(loggingService, times(1)).logInfo("Successfully completed Fetch XML batch job");
-    }
-
-    @DisplayName("Test FetchXML Exception")
+    @DisplayName("Test Migrating Exclusions Exception")
     @Test
     public void testRunException() throws JobInstanceAlreadyCompleteException,
         JobExecutionAlreadyRunningException,
         JobParametersInvalidException,
         JobRestartException {
 
-        when(jobLauncher.run(eq(fetchXmlJob), any()))
+        when(jobLauncher.run(eq(processExclusionsJob), any()))
             .thenThrow(new JobExecutionAlreadyRunningException("Test"));
 
-        var fetchXML = new FetchXML(userService,
-                                    userAuthenticationService,
-                                    CRON_USER_EMAIL,
-                                    jobLauncher,
-                                    loggingService,
-                                    false,
-                                    false,
-                                    MigrationType.FULL.name(),
-                                    fetchXmlJob);
-        fetchXML.run();
+        var migrateExclusions = new MigrateExclusions(userService,
+                                                      userAuthenticationService,
+                                                      CRON_USER_EMAIL,
+                                                      jobLauncher,
+                                                      loggingService,
+                                                      false,
+                                                      false,
+                                                      processExclusionsJob);
+        migrateExclusions.run();
 
         verify(loggingService, times(1))
-            .logError(eq("Error starting Fetch XML batch job"),
+            .logError(eq("Error starting Process Exclusions batch job"),
                       any(JobExecutionAlreadyRunningException.class));
     }
 }
