+++ src/test/java/uk/gov/hmcts/reform/preapi/tasks/migration/ProcessMigrationTest.java
@@ -1,4 +1,4 @@
-package uk.gov.hmcts.reform.preapi.tasks.batch;
+package uk.gov.hmcts.reform.preapi.tasks.migration;
 
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
@@ -19,7 +19,6 @@ import uk.gov.hmcts.reform.preapi.dto.base.BaseAppAccessDTO;
 import uk.gov.hmcts.reform.preapi.security.authentication.UserAuthentication;
 import uk.gov.hmcts.reform.preapi.security.service.UserAuthenticationService;
 import uk.gov.hmcts.reform.preapi.services.UserService;
-import uk.gov.hmcts.reform.preapi.tasks.migration.MigrateExclusions;
 
 import java.util.Optional;
 import java.util.Set;
@@ -32,13 +31,13 @@ import static org.mockito.Mockito.times;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 
-public class MigrateExclusionsTest {
+public class ProcessMigrationTest {
 
     private static UserService userService;
     private static UserAuthenticationService userAuthenticationService;
     private static JobLauncher jobLauncher;
     private static LoggingService loggingService;
-    private static Job processExclusionsJob;
+    private static Job importCsvJob;
 
     private static final String CRON_USER_EMAIL = "test@test.com";
 
@@ -59,61 +58,110 @@ public class MigrateExclusionsTest {
 
         jobLauncher = mock(JobLauncher.class);
         loggingService = mock(LoggingService.class);
-        processExclusionsJob = mock(Job.class);
+        importCsvJob = mock(Job.class);
     }
 
-    @DisplayName("Test Migrating Exclusions")
+    @DisplayName("Test Transform")
     @Test
     public void testRun() throws JobInstanceAlreadyCompleteException,
         JobExecutionAlreadyRunningException,
         JobParametersInvalidException,
         JobRestartException {
 
-        var migrateExclusions = new MigrateExclusions(userService,
-                                                      userAuthenticationService,
-                                                      CRON_USER_EMAIL,
-                                                      jobLauncher,
-                                                      loggingService,
-                                                      false,
-                                                      false,
-                                                      processExclusionsJob);
-        migrateExclusions.run();
+        var processMigration = new ProcessMigration(
+            userService,
+            userAuthenticationService,
+            CRON_USER_EMAIL,
+            jobLauncher,
+            loggingService,
+            false,
+            false,
+            MigrationType.FULL.name(),
+            importCsvJob
+        );
+        processMigration.run();
 
         ArgumentCaptor<JobParameters> jobParameters = ArgumentCaptor.forClass(JobParameters.class);
 
-        verify(jobLauncher, times(1)).run(eq(processExclusionsJob), jobParameters.capture());
+        verify(jobLauncher, times(1)).run(eq(importCsvJob), jobParameters.capture());
 
-        Assertions.assertEquals(String.valueOf(false),
-                                jobParameters.getValue().getString("debug"));
+        Assertions.assertEquals(
+            String.valueOf(false),
+            jobParameters.getValue().getString("debug")
+        );
 
-        Assertions.assertEquals(MigrationType.FULL.name(),
-                                jobParameters.getValue().getString("migrationType"));
+        Assertions.assertEquals(
+            MigrationType.FULL.name(),
+            jobParameters.getValue().getString("migrationType")
+        );
 
-        verify(loggingService, times(1)).logInfo("Successfully completed Process Exclusions batch job");
+        verify(loggingService, times(1)).logInfo("Successfully completed Transform batch job");
     }
 
-    @DisplayName("Test Migrating Exclusions Exception")
+    @DisplayName("Test Transform Second Type")
+    @Test
+    public void testRunSecondType() throws JobInstanceAlreadyCompleteException,
+        JobExecutionAlreadyRunningException,
+        JobParametersInvalidException,
+        JobRestartException {
+
+        var processMigration = new ProcessMigration(
+            userService,
+            userAuthenticationService,
+            CRON_USER_EMAIL,
+            jobLauncher,
+            loggingService,
+            false,
+            false,
+            MigrationType.DELTA.name(),
+            importCsvJob
+        );
+        processMigration.run();
+
+        ArgumentCaptor<JobParameters> jobParameters = ArgumentCaptor.forClass(JobParameters.class);
+
+        verify(jobLauncher, times(1)).run(eq(importCsvJob), jobParameters.capture());
+
+        Assertions.assertEquals(
+            String.valueOf(false),
+            jobParameters.getValue().getString("debug")
+        );
+
+        Assertions.assertEquals(
+            MigrationType.DELTA.name(),
+            jobParameters.getValue().getString("migrationType")
+        );
+
+        verify(loggingService, times(1)).logInfo("Successfully completed Transform batch job");
+    }
+
+    @DisplayName("Test ProcessMigration Exception")
     @Test
     public void testRunException() throws JobInstanceAlreadyCompleteException,
         JobExecutionAlreadyRunningException,
         JobParametersInvalidException,
         JobRestartException {
 
-        when(jobLauncher.run(eq(processExclusionsJob), any()))
+        when(jobLauncher.run(eq(importCsvJob), any()))
             .thenThrow(new JobExecutionAlreadyRunningException("Test"));
 
-        var migrateExclusions = new MigrateExclusions(userService,
-                                                      userAuthenticationService,
-                                                      CRON_USER_EMAIL,
-                                                      jobLauncher,
-                                                      loggingService,
-                                                      false,
-                                                      false,
-                                                      processExclusionsJob);
-        migrateExclusions.run();
+        var processMigration = new ProcessMigration(
+            userService,
+            userAuthenticationService,
+            CRON_USER_EMAIL,
+            jobLauncher,
+            loggingService,
+            false,
+            false,
+            MigrationType.FULL.name(),
+            importCsvJob
+        );
+        processMigration.run();
 
         verify(loggingService, times(1))
-            .logError(eq("Error starting Process Exclusions batch job"),
-                      any(JobExecutionAlreadyRunningException.class));
+            .logError(
+                eq("Error starting Transform batch job"),
+                any(JobExecutionAlreadyRunningException.class)
+            );
     }
 }
